readObjectPy/requirements.txt
functions-framework==3.1.0
flask==2.1.0
google-cloud-error-reporting==1.5.3


google-cloud-pubsub==2.13.0
google-cloud-storage==2.4.0
pandas===1.3.5; python_version == '3.7'
pandas==1.4.2; python_version >= '3.8'